usecases/UPDATE NEEDED/T1_2025_LLM_Chatbot/gpt_mental_health_bot_faiz/requirements.txt
absl-py==2.2.0
accelerate==1.5.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bert-score==0.3.13
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.4.1
dill==0.3.8
distro==1.9.0
evaluate==0.4.0
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.47
langchain-openai==0.3.9
langchain-text-splitters==0.3.7
langsmith==0.3.18
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
openai==1.68.2
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.3.0
psutil==7.0.0
pyarrow==19.0.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
responses==0.18.0
rouge_score==0.1.2
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
sympy==1.13.3
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.15.2
torch==2.2.2
tqdm==4.67.1
transformers==4.36.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
flask
